nonsens add (#3)
* nonsens add

* localname update

* deployop update

* validatechecks

* change deploy check time

* cmt

* add instance deploy

* annotation fix

* add pull secret

* add dependencies

* cr-add test from fork

* docupdate

* chng namespace timeout

+++ .github/scripts/validate-deploy.sh
@@ -8,9 +8,9 @@ NAMESPACE=$(cat .namespace)
 BRANCH="main"
 SERVER_NAME="default"
 TYPE="base"
-LAYER="3-applications"
+LAYER="2-services"
 
-COMPONENT_NAME="turbonomic"
+COMPONENT_NAME="turbo"
 
 mkdir -p .testrepo
 
@@ -28,14 +28,6 @@ fi
 echo "Printing argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml"
 cat "argocd/${LAYER}/cluster/${SERVER_NAME}/${TYPE}/${NAMESPACE}-${COMPONENT_NAME}.yaml"
 
-#if [[ ! -f "payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/xl.yaml" ]]; then
-#  echo "Application values not found - payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/xl.yaml"
-#  exit 1
-#fi
-
-#echo "Printing payload/${LAYER}/namespace/${NAMESPACE}/${COMPONENT_NAME}/xl.yaml"
-#cat "payload/${LAYER}namespace/${NAMESPACE}/${COMPONENT_NAME}/xl.yaml"
-
 count=0
 until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $count -eq 20 ]]; do
   echo "Waiting for namespace: ${NAMESPACE}"
@@ -43,7 +35,7 @@ until kubectl get namespace "${NAMESPACE}" 1> /dev/null 2> /dev/null || [[ $coun
   sleep 15
 done
 
-if [[ $count -eq 20 ]]; then
+if [[ $count -eq 30 ]]; then
   echo "Timed out waiting for namespace: ${NAMESPACE}"
   exit 1
 else
@@ -51,9 +43,24 @@ else
   sleep 30
 fi
 
+### validation logic checks ####
 
-### add more validation logic when done ####
-#kubectl rollout status "deployment/${DEPLOYMENT}" -n "${NAMESPACE}" || exit 1
+#wait for deployment
+sleep 8m
+
+count=0
+until kubectl rollout status deployment/t8c-operator -n "${NAMESPACE}" || [[ $count -eq 4 ]]; do
+  echo "Waiting for turbo operator rollout to deploy"
+  count=$((count + 1))
+  sleep 15
+done
+
+if [[ $count -eq 4 ]]; then
+  echo "Timed out waiting for turbo operator to deploy"
+  exit 1
+fi
+
 
 cd ..
 rm -rf .testrepo
+
